Create 3 fragments.
These fragments will change in the container on the screen when you click on the corresponding menu items. They should also be added to the FragmentManager backstack.

Main - opens by default when the application starts. (added to the container). Add the markup to your discretion.

Settings - opens by clicking on the menu item Settings. On the layout are RadioButtons - three items - Google, Yandex, Bing. Save the selected item in SharedPreferences. Extract from SharedPreferences when the screen is launched.

Search - opens by clicking on the Search item. The markup is EditText and the Search button. When you click on the button - the browser opens with the search of the entered text in the search engine selected in the settings screen.

When you click Exit, the application closes.